Fix type def
